+++ E3Next/DiscordBot/DiscordMessageRequest.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace E3Core.DiscordBot
+{
+    public class DiscordMessageRequest
+    {
+        public string content { get; set; }
+    }
+}
