+++ E3Next/E3Next.csproj
@@ -113,14 +113,17 @@
     <Compile Include="Data\SpellRequest.cs" />
     <Compile Include="Data\SpellTimer.cs" />
     <Compile Include="Data\Zone.cs" />
+    <Compile Include="DiscordBot\DiscordMessage.cs" />
+    <Compile Include="DiscordBot\DiscordMessageRequest.cs" />
+    <Compile Include="DiscordBot\DiscordUser.cs" />
+    <Compile Include="DiscordBot\GuildMember.cs" />
     <Compile Include="Processors\BaseProcessor.cs" />
     <Compile Include="Processors\Bots.cs" />
     <Compile Include="Processors\Burns.cs" />
     <Compile Include="Processors\Charm.cs" />
     <Compile Include="Processors\ClearXTargets.cs" />
     <Compile Include="Processors\DebuffDot.cs" />
-    <Compile Include="Processors\DiscordClient2.cs" />
-    <Compile Include="Processors\DiscordClient.cs" />
+    <Compile Include="DiscordBot\DiscordClient.cs" />
     <Compile Include="Processors\Dispel.cs" />
     <Compile Include="Processors\GiveMe.cs" />
     <Compile Include="Processors\Inventory.cs" />
